+++ tests/test_cluster.py
@@ -29,7 +29,7 @@ class graph:
         g = ig.Graph.Preference(
             n_nodes, type_dist.tolist(), pref, attribute="type"
         )
-        ground_truth = ig.VertexClustering(g, g.vs["type"])
+        ground_truth = g.vs["type"]
 
         return (g, ground_truth)